+++ dst/app/routes/reservations.js
@@ -13,7 +13,6 @@ exports.cancelReservation = void 0;
 /**
  * 予約ルーター
  */
-const chevre = require("@chevre/api-nodejs-client");
 const cinerinoapi = require("@cinerino/api-nodejs-client");
 const ttts = require("@tokyotower/domain");
 const express = require("express");
@@ -317,7 +316,10 @@ reservationsRouter.delete('/:id/checkins/:when', permitScopes_1.default(['reserv
  */
 reservationsRouter.put('/:id/cancel', permitScopes_1.default(['admin']), validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
-        yield cancelReservation({ id: req.params.id })({
+        yield cancelReservation({
+            req: req,
+            id: req.params.id
+        })({
             reservation: new ttts.repository.Reservation(mongoose.connection),
             task: new ttts.repository.Task(mongoose.connection),
             project: new ttts.repository.Project(mongoose.connection)
@@ -335,16 +337,20 @@ exports.default = reservationsRouter;
  */
 function cancelReservation(params) {
     return (repos) => __awaiter(this, void 0, void 0, function* () {
-        var _a, _b;
-        const projectDetails = yield repos.project.findById({ id: project.id });
-        if (typeof ((_b = (_a = projectDetails.settings) === null || _a === void 0 ? void 0 : _a.chevre) === null || _b === void 0 ? void 0 : _b.endpoint) !== 'string') {
-            throw new ttts.factory.errors.ServiceUnavailable('Project settings not satisfied');
-        }
-        const cancelReservationService = new chevre.service.transaction.CancelReservation({
-            endpoint: projectDetails.settings.chevre.endpoint,
-            auth: cinerinoAuthClient
+        const authClient = new cinerinoapi.auth.ClientCredentials({
+            domain: '',
+            clientId: '',
+            clientSecret: '',
+            scopes: [],
+            state: ''
         });
-        yield cancelReservationService.startAndConfirm({
+        authClient.setCredentials({ access_token: params.req.accessToken });
+        const reservationService = new cinerinoapi.service.Reservation({
+            auth: authClient,
+            endpoint: process.env.CINERINO_API_ENDPOINT,
+            project: { id: project.id }
+        });
+        yield reservationService.cancel({
             project: project,
             typeOf: ttts.factory.chevre.transactionType.CancelReservation,
             agent: {
@@ -360,7 +366,6 @@ function cancelReservation(params) {
                 .add(1, 'minutes')
                 .toDate()
         });
-        // await cancelReservationService.confirm({ id: cancelReservationTransaction.id });
         // 東京タワーDB側の予約もステータス変更
         yield repos.reservation.cancel({ id: params.id });
     });