+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 
 - イベントインポート処理をイベント変更時処理へ移行
+- Chevre予約取消をCinerinoで処理するように変更
 
 ### Deprecated